Rename modules(make it more comfortable to debug)
