+++ sources/projects/AnalysisServer/src/main/java/org/civilla/AnalysisServer/AnalysisServer.java
@@ -1,11 +1,11 @@
-package org.civilla.PoliceControlServer;
+package org.civilla.AnalysisServer;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Main {
+public class AnalysisServer {
 	public static void main(String[] args) {
-		SpringApplication.run(Main.class, args);
+		SpringApplication.run(AnalysisServer.class, args);
 	}
 }
