+++ sources/projects/PoliceControlServer/src/main/java/org/civilla/PoliceControlServer/PoliceControlServer.java
@@ -1,11 +1,11 @@
-package org.civilla.AnalysisServer;
+package org.civilla.PoliceControlServer;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Main {
+public class PoliceControlServer {
 	public static void main(String[] args) {
-		SpringApplication.run(Main.class, args);
+		SpringApplication.run(PoliceControlServer.class, args);
 	}
 }
